requirements.txt
pandas
requests
mysql-connector
fastapi
uvicorn
pytest
httpx
